\name{accum}
\alias{accum}
\title{accum}
\description{utility function to calculate cumulative statistics of multiple simulations of the same parameter set. It suppose to be called from \code{ddply} function of \code{plyr} package. }
\usage{accum(.x, templ)}
\arguments{
  \item{.x}{
}
  \item{templ}{
}
}



\author{Anatoly Sorokin}





