\name{makeIGraph}
\alias{makeIGraph}
\title{makeIGraph}
\description{function converts kappa string, defining the complex structure into iGraph graph}
\usage{makeIGraph(kappa, num = 1)}
\arguments{
  \item{kappa}{complex definition in kappa language}
  \item{num}{number of complex instances in the mixture}
}

\value{simplified iGraph graph of the complex}

\author{Anatoly Sorokin}





