\name{makeSIF}
\alias{makeSIF}
\title{makeSIF}
\description{function creates SIF (Simple Interaction Format) from kappa string}
\usage{makeSIF(kappa)}
\arguments{
  \item{kappa}{kappa string defining the complex structure}
}

\value{simplified iGraph graph of the complex}

\author{Anatoly Sorokin}





