\name{makeSiteGraph}
\alias{makeSiteGraph}
\title{makeSiteGraph}
\description{function to create igraph representation of kappa string with sites being separate nodes}
\usage{makeSiteGraph(kp)}
\arguments{
  \item{kp}{kappa string to convert}
}

\value{site graph in igraph }

\author{Anatoly Sorokin}





