\name{plot.point.sensitivity}
\alias{plot.point.sensitivity}
\title{plot point sensitivity}

\usage{\method{plot}{point.sensitivity}(kasens, npar = 30, conf.value = 0.05)}
\arguments{
  \item{kasens}{
}
  \item{npar}{
}
  \item{conf.value}{
}
}



\author{Anatoly Sorokin}





