\name{read.observables.default}
\alias{read.observables.default}
\title{read observables default}
\description{function to read and parse KaSim simulation results}
\usage{\method{read.observables}{default}(file, ...)}
\arguments{
  \item{file}{name of the observable file}
  \item{\dots}{other parameters}
}

\value{data frame of observables}

\author{Anatoly Sorokin}





