\name{read.observables}
\alias{read.observables}
\title{read observables}
\description{function to read and parse KaSim simulation results}
\usage{read.observables(object, ...)}
\arguments{
  \item{object}{object to read from}
  \item{\dots}{other parameters}
}



\author{Anatoly Sorokin}





