\name{read.snapshot.kproject}
\alias{read.snapshot.kproject}
\title{read snapshot kproject}
\description{function to read and parse KaSim simulation results}
\usage{read.snapshot.kproject(kproject, dir)}
\arguments{
  \item{kproject}{project to handle}
  \item{dir}{name of the folder with project simulation results}
}



\author{Anatoly Sorokin}





