\name{readFiles}
\alias{readFiles}
\title{readFiles}
\description{Read file contents into character list}
\usage{readFiles(files)}
\arguments{
  \item{files}{vector of file names to read}
}

\value{list of contents of the files. Each element of the list named after a file it contains.}

\author{Anatoly Sorokin}





