\name{replaceProperty}
\alias{replaceProperty}
\title{replaceProperty}
\description{Replace named parameter with project property}
\usage{replaceProperty(kproject, pname, lines)}
\arguments{
  \item{kproject}{project to take property from}
  \item{pname}{name of the property to substitute with value}
  \item{lines}{string vector to make substitution in}
}



\author{Anatoly Sorokin}





