\name{rkappa-package}
\alias{rkappa-package}
\alias{rkappa}
\docType{package}
\title{Kappa simulation Project}

\description{The package provides a infrastructure required to simulate kappa model with various parameter sets and analyse results of such simulations.}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{rkappa}
\packageIndices{rkappa}

\author{Anatoly Sorokin}

Maintainer: \packageMaintainer{rkappa}


\keyword{ package }


