\name{summary.kproject}
\alias{summary.kproject}
\title{summary kproject}
\description{Kproject class summary method}
\usage{\method{summary}{kproject}(object, ...)}
\arguments{
  \item{object}{object to create summary for}
  \item{\dots}{
}
}

\value{kproject.summary object}

\author{Anatoly Sorokin}





