\name{write.kproject}
\alias{write.kproject}
\title{write kproject}
\description{Write content of the \code{kproject} to the folder}
\usage{write.kproject(kproject, projectdir = kproject$name)}
\arguments{
  \item{kproject}{object to write}
  \item{projectdir}{optional new destination for the writing}
}



\author{Anatoly Sorokin}





